ported stuff from mainline Linux to here. preparing from the next patch
